Chnage module from stash to bitbucket server
Prior to this commit, this profile was for setting up stash using
puppet/stash but now atlassian changed the name of stash to
bitbucket server and changed the installer a bit.

After this commit, this profile using thewired/bitbucket module
and made some changes to the profile to meet the dependency
requirements for bitbucket server.  These include java 8 and
git 1.8+